+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - bug fixes
   - meter's properties not accessible (overridden with fader's properties)
   - `--read-only` option not properly working
+  - `keyboard`: dynamic properties `target`, `precision` & `preArgs` not being updated dynamically
 - widgets
   - formulas: added `indexOf` function
   - **deprecated** `multifader` / `multipush` / `multitoggle` in favor of the generic `matrix` widget